chore: Updated agent integration tests to node:test
